+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2021-01-09 - Edu
+## changed
+- **utils/error_handler.c** replace usleep with bash sleep through execve 
+
 ## 2021-01-05 - Edu
 ## Added
 - new folder **forks** to storage related files